requirements.txt
flask
flask-restplus
Flask-SQLAlchemy
pytest